Add some refactoring
