+++ main.rb
@@ -1,4 +1,4 @@
-require_relative 'app.rb'
+require_relative 'app'
 
 def main
   app = App.new
